Fix panic when exit code does not have a cause
Fixes: https://github.com/homeport/dyff/issues/448

Use `err.Error()` of `ExitCode` interface since this has a nil check.

Move `ExitCode` into its own source file.

Fix TYPO in comment about error.
